+++ example/javascriptlib/basic/1-simple/build.mill
@@ -1,10 +1,10 @@
 package build
 
-import mill._, javascriptlib._, javascriptlib.testmods._
+import mill._, javascriptlib._
 
-object foo extends JestModule {
+object foo extends TypeScriptModule {
   def npmDeps = Seq("immutable@4.3.7")
-
+  object test extends TypeScriptTests with TestModule.Jest
 }
 
 // Documentation for mill.example.javascriptlib
@@ -15,6 +15,9 @@ Hello James Bond Professor
 
 > mill foo.test
 PASS .../foo.test.ts
+...
+Test Suites:...1 passed, 1 total...
+Tests:...3 passed, 3 total...
 ...
 
 > mill show foo.bundle
