+++ example/javascriptlib/basic/1-simple/foo/test/src/foo/foo.test.ts
@@ -1,5 +1,5 @@
-import generateUser from "../src/foo";
-import {compareObject} from 'bar/test/utils/bar.tests.utils'
+import {generateUser, defaultRoles} from "foo/foo";
+import {Map} from 'node_modules/immutable';
 
 // Define the type roles object
 type RoleKeys = "admin" | "user";
@@ -7,15 +7,11 @@ type Roles = {
     [key in RoleKeys]: string;
 };
 
-// Mock the defaultRoles.get method
-jest.mock("bar/bar", () => ({
-    defaultRoles: {
-        get: jest.fn((role: string, defaultValue: string) => {
-            const roles: Roles = {"admin": "Administrator", "user": "User"};
-            return roles[role as RoleKeys] || defaultValue;
-        }),
-    },
-}));
+// Mock `defaultRoles` as a global variable for testing
+const mockDefaultRoles = Map<string, string>({
+    admin: "Administrator",
+    user: "User",
+});
 
 describe("generateUser function", () => {
     beforeAll(() => {
@@ -27,35 +23,42 @@ describe("generateUser function", () => {
     });
 
     test("should generate a user with all specified fields", () => {
+        // Override the `defaultRoles` map for testing
+        (defaultRoles as any).get = mockDefaultRoles.get.bind(mockDefaultRoles);
+
         const args = ["John", "Doe", "admin"];
         const user = generateUser(args);
 
-        expect(compareObject({...user}, {
+        expect(user).toEqual({
             firstName: "John",
             lastName: "Doe",
             role: "Administrator",
-        })).toBeTruthy()
+        });
     });
 
     test("should default lastName and role when they are not provided", () => {
+        (defaultRoles as any).get = mockDefaultRoles.get.bind(mockDefaultRoles);
+
         const args = ["Jane"];
         const user = generateUser(args);
 
-        expect(compareObject(user, {
+        expect(user).toEqual({
             firstName: "Jane",
             lastName: "unknown",
             role: "",
-        })).toBeTruthy()
+        });
     });
 
     test("should default all fields when args is empty", () => {
+        (defaultRoles as any).get = mockDefaultRoles.get.bind(mockDefaultRoles);
+
         const args: string[] = [];
         const user = generateUser(args);
 
-        expect(compareObject(user, {
+        expect(user).toEqual({
             firstName: "unknown",
             lastName: "unknown",
             role: "",
-        })).toBeTruthy()
+        });
     });
 });
