+++ example/javascriptlib/basic/3-custom-build-logic/foo/test/src/foo/foo.test.ts
@@ -1,6 +1,6 @@
 import * as fs from 'fs';
 import * as path from 'path';
-import Foo from '../src/foo';
+import Foo from 'foo/foo';
 
 // Mock the 'fs' module
 jest.mock('fs');
