+++ example/javascriptlib/testing/1-test-suite/build.mill
@@ -1,10 +1,14 @@
 package build
 
-import mill._, javascriptlib._, javascriptlib.testmods._
+import mill._, javascriptlib._
 
-object bar extends JestModule
+object bar extends TypeScriptModule {
+  object test extends TypeScriptTests with TestModule.Jest
+}
 
-object foo extends MochaModule
+object foo extends TypeScriptModule {
+  object test extends TypeScriptTests with TestModule.Mocha
+}
 
 // Documentation for mill.example.javascriptlib
 // This build defines two modules bar and foo with test suites configured to use
@@ -13,10 +17,14 @@ object foo extends MochaModule
 /** Usage
 
 > mill foo.test
-...Calculator
+...
+...4 passing...
 ...
 
 > mill bar.test
-PASS .../calculator.test.ts
+...Calculator
+...
+Test Suites:...1 passed, 1 total...
+Tests:...4 passed, 4 total...
 ...
 */
