+++ example/javascriptlib/testing/1-test-suite/foo/src/calculator.ts
@@ -3,14 +3,6 @@ export class Calculator {
         return a + b;
     }
 
-    subtract(a: number, b: number): number {
-        return a - b;
-    }
-
-    multiply(a: number, b: number): number {
-        return a * b;
-    }
-
     divide(a: number, b: number): number {
         if (b === 0) {
             throw new Error("Division by zero is not allowed");
