+++ example/javascriptlib/testing/1-test-suite/jest.config.ts
@@ -1,13 +1,24 @@
+// @ts-nocheck
+import {pathsToModuleNameMapper} from 'node_modules/ts-jest';
+import {compilerOptions} from './tsconfig.json'; // this is a generated file.
+
+// Remove unwanted keys
+const moduleDeps = {...compilerOptions.paths};
+delete moduleDeps['*'];
+delete moduleDeps['typeRoots'];
+
+
 export default {
     preset: 'ts-jest',
     testEnvironment: 'node',
     testMatch: [
-        '<rootDir>/**/test/**/*.test.ts',
-        '<rootDir>/**/test/**/*.test.js',
+        '<rootDir>/**/**/**/*.test.ts',
+        '<rootDir>/**/**/**/*.test.js',
     ],
     transform: {
-        '^.+\\.(ts|tsx)$': ['ts-jest', { tsconfig: 'tsconfig.json' }],
+        '^.+\\.(ts|tsx)$': ['ts-jest', {tsconfig: 'tsconfig.json'}],
         '^.+\\.(js|jsx)$': 'babel-jest', // Use babel-jest for JS/JSX files
     },
-    moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node']
+    moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
+    moduleNameMapper: pathsToModuleNameMapper(moduleDeps) // use absolute paths generated in tsconfig.
 };
