+++ example/javascriptlib/testing/2-test-deps/foo/test/src/foo/foo.test.ts
@@ -1,5 +1,5 @@
-import {generateUser, defaultRoles} from "../src/foo";
-import {Map} from 'node_modules/immutable';
+import generateUser from "foo/foo";
+import {compareObject} from 'bar/test/utils/bar.tests.utils'
 
 // Define the type roles object
 type RoleKeys = "admin" | "user";
@@ -7,11 +7,15 @@ type Roles = {
     [key in RoleKeys]: string;
 };
 
-// Mock `defaultRoles` as a global variable for testing
-const mockDefaultRoles = Map<string, string>({
-    admin: "Administrator",
-    user: "User",
-});
+// Mock the defaultRoles.get method
+jest.mock("bar/bar", () => ({
+    defaultRoles: {
+        get: jest.fn((role: string, defaultValue: string) => {
+            const roles: Roles = {"admin": "Administrator", "user": "User"};
+            return roles[role as RoleKeys] || defaultValue;
+        }),
+    },
+}));
 
 describe("generateUser function", () => {
     beforeAll(() => {
@@ -23,42 +27,35 @@ describe("generateUser function", () => {
     });
 
     test("should generate a user with all specified fields", () => {
-        // Override the `defaultRoles` map for testing
-        (defaultRoles as any).get = mockDefaultRoles.get.bind(mockDefaultRoles);
-
         const args = ["John", "Doe", "admin"];
         const user = generateUser(args);
 
-        expect(user).toEqual({
+        expect(compareObject({...user}, {
             firstName: "John",
             lastName: "Doe",
             role: "Administrator",
-        });
+        })).toBeTruthy()
     });
 
     test("should default lastName and role when they are not provided", () => {
-        (defaultRoles as any).get = mockDefaultRoles.get.bind(mockDefaultRoles);
-
         const args = ["Jane"];
         const user = generateUser(args);
 
-        expect(user).toEqual({
+        expect(compareObject(user, {
             firstName: "Jane",
             lastName: "unknown",
             role: "",
-        });
+        })).toBeTruthy()
     });
 
     test("should default all fields when args is empty", () => {
-        (defaultRoles as any).get = mockDefaultRoles.get.bind(mockDefaultRoles);
-
         const args: string[] = [];
         const user = generateUser(args);
 
-        expect(user).toEqual({
+        expect(compareObject(user, {
             firstName: "unknown",
             lastName: "unknown",
             role: "",
-        });
+        })).toBeTruthy()
     });
 });
