+++ example/javascriptlib/testing/2-test-deps/jest.config.ts
@@ -7,12 +7,13 @@ const moduleDeps = {...compilerOptions.paths};
 delete moduleDeps['*'];
 delete moduleDeps['typeRoots'];
 
+
 export default {
     preset: 'ts-jest',
     testEnvironment: 'node',
     testMatch: [
-        '<rootDir>/**/test/**/*.test.ts',
-        '<rootDir>/**/test/**/*.test.js',
+        '<rootDir>/**/**/**/*.test.ts',
+        '<rootDir>/**/**/**/*.test.js',
     ],
     transform: {
         '^.+\\.(ts|tsx)$': ['ts-jest', {tsconfig: 'tsconfig.json'}],
